preview listing changes
