+++ src/views/Exercise/Details/Summary/Edit.vue
@@ -26,12 +26,12 @@
           required
         />
 
-        <Checkbox
+        <!-- <Checkbox
           id="invite-only"
           v-model="formData.inviteOnly"
           label="Invite only exercise"
           hint="When this is checked, only invited candidates will be able to apply to this exercise."
-        />
+        /> -->
 
         <DateInput
           id="estimated-launch-date"
@@ -97,6 +97,15 @@
           type="url"
         />
 
+        <ListingPreview
+          class="govuk-!-margin-bottom-4"
+          :exercise="{...exercise, ...formData }"
+        />
+        <DetailPreview
+          v-if="advertType !== 'listing'"
+          :exercise="{...exercise, ...formData }"
+        />
+
         <button class="govuk-button">
           Save and continue
         </button>
@@ -116,6 +125,9 @@ import { exerciseAdvertTypes } from '@/helpers/exerciseHelper';
 import { ADVERT_TYPES } from '@/helpers/constants';
 import BackLink from '@jac-uk/jac-kit/draftComponents/BackLink';
 import Checkbox from '@jac-uk/jac-kit/draftComponents/Form/Checkbox';
+import ListingPreview from '@/components/Previews/ListingPreview.vue';
+import DetailPreview from '@/components/Previews/DetailPreview.vue';
+import exerciseMixin from '@/views/Exercise/exerciseMixin.js';
 
 export default {
   name: 'SummaryEdit',
@@ -127,8 +139,11 @@ export default {
     Select,
     BackLink,
     Checkbox,
+    ListingPreview,
+    DetailPreview,
   },
   extends: Form,
+  mixins: [exerciseMixin],
   data() {
     const defaults = {
       name: null,
@@ -148,6 +163,9 @@ export default {
     };
   },
   computed: {
+    exercise() {
+      return this.$store.state.exerciseDocument.record;
+    },
     launchDate: {
       get() {
         return this.parseDate(this.formData.estimatedLaunchDate);
